+++ doc/changelog.rst
@@ -1,6 +1,13 @@
 Changelog
 =========
 
+Version 0.7.13
+--------------
+
+* Run constructor now accepts an empty list of samples. Moreover, it also
+  accepts ``int`` and ``long`` number types for warmup sample values, not only
+  ``float``.
+
 Version 0.7.12 (2016-09-30)
 ---------------------------